chg more coordinates to use double not float
git-svn-id: http://vtp.googlecode.com/svn/trunk@1276 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/BExtractor/BExtractorDoc.cpp
@@ -247,12 +247,12 @@ int BExtractorDoc::Fill(CDib *bm, int x, int y, byte old_v, byte new_v,
 	return size;
 }
 
-float BExtractorDoc::i_UTMx(int ix)
+double BExtractorDoc::i_UTMx(int ix)
 {
 	return (ix * m_pImage->m_xMetersPerPixel + m_pImage->m_xUTMoffset);
 }
 
-float BExtractorDoc::i_UTMy(int iy)
+double BExtractorDoc::i_UTMy(int iy)
 {
 	return ((m_pImage->m_PixelSize.y - 1 - iy) * m_pImage->m_yMetersPerPixel
 		+ m_pImage->m_yUTMoffset);
